+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 - Added a new validation check for position shift into a pull-off.
 - Fixed false positive in validation for anchor position in relation to chord fingering with chords that use the thumb.
 - A better error message is shown when loading an "EXT" vocals file fails.
+- Phrase level file is now created also when a PSARC is imported.
 - Minor UI improvements.
 - Minor optimizations to package generation.
 - Updated the UI framework to new minor version.