+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - A cancel button is now shown when in quick edit "mode".
 - Added possibility to launch the custom font generator from the UI.
 - Added an option for the naming of the base tone key created automatically.
+- Added an option to compare phrase level counts to previous version for test builds.
 - Added a new validation check for hammer-ons/pull-offs that are on the same fret as the previous note on the same string.
 - Added a new validation check for a finger changing during a slide.
 - Added a new validation check for position shift into a pull-off.